pyproject.toml
[project]
name = "fproject"
version = "0.1.0"
description = ""
authors = [
    {name = "None"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=5.1.5,<6.0.0)",
    "redis (>=5.2.1,<6.0.0)",
    "djangorestframework (>=3.15.2,<4.0.0)",
    "django-redis (>=5.4.0,<6.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
